inference-endpoint/requirements.txt
numpy==1.20.1
Pillow==8.1.0
torch==1.7.1+cpu
torch-model-archiver==0.3.0
torchserve==0.3.0
torchvision==0.8.2+cpu